requirements.txt
tensorflow==2.12.0
numpy==1.23.5
scipy==1.13.0
matplotlib>=3.0
pandas>=1.3.0
opencv-python
scikit-learn>=1.0
Pillow
streamlit
